refactor: party cards components
